Initial Push (p1)
Most of this is some code I had worked on before hand

As of now I am running into an issue with re-rendering the physical array of Todos onto the page, instead a random {Object: object} pops up every time, not sure why.  Probably something to do with TodoForm.js?  Working on it